% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_liste_grille.R
\name{gg_liste_grille}
\alias{gg_liste_grille}
\title{Assembler des graphiques comparant les sous-lots}
\usage{
gg_liste_grille(
  liste_graphiques,
  n_colonnes,
  n_lignes,
  titre_axe_x = "",
  titre_axe_y = ""
)
}
\arguments{
\item{liste_graphiques}{Liste contenant les graphiques à assembler en grille.}

\item{n_colonnes}{Entier. Nombre de colonnes de la grille.}

\item{n_lignes}{Entier. Nombre de lignes de la grille.}

\item{titre_axe_x}{Caractère. Titre de l'axe des abscisses, commun aux graphiques de la grille.}

\item{titre_axe_y}{Caractère. Titre de l'axe des ordonnées, commun aux graphiques de la grille.}
}
\value{
Le graphique ggplot assemblé.
}
\description{
S'applique typiquement sur des graphiques issus de qta_densite_comp_distri_lot().
}
\examples{
\dontrun{
# Avec 12 graphiques à assembler
gg_liste_grille(liste_graphiques = graphiques, ncol = 3, nrow = 4)
}
}
