% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qta_densite_comp_distri_lot.R
\name{qta_densite_comp_distri_lot}
\alias{qta_densite_comp_distri_lot}
\title{Grapher le density plot de deux sous-lots}
\usage{
qta_densite_comp_distri_lot(df, id_lot, variable)
}
\arguments{
\item{df}{Dataframe de données avec une variable "lop_id" et le nom de la variable choisie
à représenter.}

\item{id_lot}{Entier. Numéro du lot.}

\item{variable}{Caractère. Nom de la variable à représenter, entre guillemets.}
}
\value{
Le graphique ggplot pour ce lot et cette variable
}
\description{
S'applique typiquement pour comparer les distributions des longueurs entre individus
    mesurés ert "dégroupés".
}
\examples{
\dontrun{
qta_densite_comp_distri_lot (df = df, id_lot = 3156360, variable = "mei_taille")
}
}
