% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qta_gg_densite_comp_distri_lot_grille.R
\name{qta_gg_densite_comp_distri_lot_grille}
\alias{qta_gg_densite_comp_distri_lot_grille}
\title{Assembler des graphiques comparant les sous-lots}
\usage{
qta_gg_densite_comp_distri_lot_grille(liste_graphiques, ncol, nrow)
}
\arguments{
\item{liste_graphiques}{Liste contenant les graphiques à assembler en grille.}

\item{ncol}{Entier. Nombre de colonnes de la grille.}

\item{nrow}{Entier. Nombre de lignes de la grille.}
}
\value{
Le graphique ggplot assemblé
}
\description{
S'applique typiquement sur des graphiques issus de qta_densite_comp_distri_lot().
}
\examples{
\dontrun{
# Avec 12 graphiques à assembler
qta_gg_densite_comp_distri_lot_grille(liste_graphiques = graphiques, ncol = 3, nrow = 4)
}
}
