% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qta_gg_histo_comp_degroupage.R
\name{qta_gg_histo_comp_degroupage}
\alias{qta_gg_histo_comp_degroupage}
\title{Grapher l'histogramme comparé des deux méthodes de dégroupage}
\usage{
qta_gg_histo_comp_degroupage(
  df,
  ope,
  espece,
  x_var = "mei_taille",
  groupe_var = "degroupage",
  bins = 20
)
}
\arguments{
\item{df}{Dataframe de données avec les variables "ope_id", "esp_nom_commun" issues des tables Aspe
ainsi que la variable choisie
à représenter.}

\item{ope}{Entier. Numéro de l'opération.}

\item{espece}{Caractère. Nom commun de l'espèce à représenter.}

\item{x_var}{Caractère. Nom de la variable dont la distribution sera représentée, entre guillemets.}

\item{groupe_var}{Caractère. Nom de la variable de groupage, entre guillemets.}

\item{bins}{Entier. Nombre d'intervalles de l'histogramme (20 par défaut).}
}
\value{
L'histogramme ggplot de x_var pour cette espèce, lors de cette opération.
}
\description{
S'applique typiquement pour comparer les distributions des longueurs entre individus
    mesurés ert "dégroupés".
}
\examples{
\dontrun{
qta_gg_histo_comp_degroupage (df = df, id_lot = 3156360, variable = "mei_taille")
}
}
