% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qta_tests_1lot_1var.R
\name{qta_tests_1lot_1var}
\alias{qta_tests_1lot_1var}
\title{Tester sur un lot et une variable si les distributions diffèrent}
\usage{
qta_tests_1lot_1var(df, lop_id, variable)
}
\arguments{
\item{df}{Dataframe de données avec une variable "lop_id" et le nom de la variable choisie
à représenter.}

\item{lop_id}{Entier. Numéro du lot.}

\item{variable}{Caractère. Nom de la variable à représenter, entre guillemets.}
}
\value{
Un dataframe avec une ligne contenant les résultats des tests.
}
\description{
Les tests sont Kolmogorov-Smirnov (comparaison des distributions), Student (comparaison des moyennes)
    et F-test (comparaison des variances).
}
\examples{
\dontrun{
tests <- qta_tests_1lot_1var(df = df, lop_id = 3156360, variable = "mei_taille")
}
}
