% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qta_tests_distrib_sl.R
\name{qta_tests_distrib_sl}
\alias{qta_tests_distrib_sl}
\title{Tester l'équivalence des distributions entre individus mesurés et dégroupées. Peut servir à
    comparer deux méthodes de dégroupage.}
\usage{
qta_tests_distrib_sl(
  df,
  lots = NA,
  var_taille1,
  var_taille2 = NA,
  var_taille3 = NA,
  sig = TRUE
)
}
\arguments{
\item{df}{Dataframe de données avec une variable "lop_id" et le nom de la (des) variable(s)
choisie(s) à représenter.}

\item{lots}{Vecteur entier. Numéros des lots (lop_id).}

\item{var_taille1}{Caractère. Nom de la variable à représenter, entre guillemets.}

\item{var_taille2}{Caractère. Nom de variable à représenter, entre guillemets (si deux méthodes
de dégroupage à comparer).}

\item{var_taille3}{Caractère. Nom de variable à représenter, entre guillemets (si trois méthodes
de dégroupage à comparer).}

\item{sig}{Booléen. Si sig = TRUE (défaut), le tableau contient les niveaux de significativité
des tests. Si sig = FALSE, il contient les p-values.}
}
\value{
Un dataframe avec une ligne par lot, contenant les résultats des tests.
}
\description{
Les tests sont Kolmogorov-Smirnov (comparaison des distributions), Student (comparaison des moyennes)
    et F-test (comparaison des variances).
}
\examples{
\dontrun{
tests <- qta_tests_distrib_sl(df = df, lots = 3156360, var_taille1 = "mei_taille")
}
}
