% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtp_seuils.R
\name{qtp_seuils}
\alias{qtp_seuils}
\title{Calculer les seuils}
\usage{
qtp_seuils(df, var_taxon, var_a_tester = mei_taille, seuil_densite = 0.01)
}
\arguments{
\item{df}{Dataframe contenant les données.}

\item{var_taxon}{Variable contenant les identifiants des taxons (codes espèces, noms scientifiques ...).}

\item{var_a_tester}{Variable dont la distribution est examinée, par exemple mei_taille ou mei_poids.}

\item{seuil_densite}{Numérique. Seuil de densité de probabilité au sens
de ggplot2::geom_density(). Par défaut il est de 1\%.}
}
\value{
Un dataframe contenant les seuils haut et bas, c'est-à-dire les tailles ou
    poids mini et maxi à retenir au seuil choisi.
}
\description{
Calculer les seuils
}
\examples{
\dontrun{
qtp_seuils (df = tp_data,
            var_taxon = esp_code_alternatif,
            var_a_tester = mei_taille,
            seuil_densite = 0.001)
}
}
